requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
jinja2==3.1.2
crewai==0.11.0
pydantic==2.11.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2 